+++ src/main/kotlin/org/balhom/currencyprofilesapi/modules/currencyprofiles/infrastructure/producers/data/CurrencyProfileChangeEvent.kt
@@ -1,14 +1,15 @@
 package org.balhom.currencyprofilesapi.modules.currencyprofiles.infrastructure.producers.data
 
 import org.balhom.currencyprofilesapi.modules.currencyprofiles.domain.models.CurrencyProfileSharedUser
+import java.math.BigDecimal
 import java.util.UUID
 
 data class CurrencyProfileChangeEvent(
     val action: String,
     val id: UUID,
-    val balance: Double,
-    val monthlyGoal: Double,
-    val yearlyGoal: Double,
+    val balance: BigDecimal,
+    val monthlyGoal: BigDecimal,
+    val yearlyGoal: BigDecimal,
     val ownerId: UUID,
     val sharedUsers: MutableList<CurrencyProfileSharedUser>,
 )
