+++ src/test/kotlin/org/balhom/currencyprofilesapi/modules/currencyprofiles/domain/models/MockCurrencyProfileFactory.kt
@@ -3,11 +3,11 @@ package org.balhom.currencyprofilesapi.modules.currencyprofiles.domain.models
 import org.balhom.currencyprofilesapi.common.data.enums.CurrencyCodeEnum
 import org.balhom.currencyprofilesapi.common.data.models.MockAuditableDataFactory
 import org.balhom.currencyprofilesapi.common.data.models.MockFileReferenceDataFactory
+import org.balhom.currencyprofilesapi.common.utils.TestDataUtils.Companion.randomBigDecimal
 import org.balhom.currencyprofilesapi.common.utils.TestDataUtils.Companion.randomBool
-import org.balhom.currencyprofilesapi.common.utils.TestDataUtils.Companion.randomDouble
 import org.balhom.currencyprofilesapi.common.utils.TestDataUtils.Companion.randomPastDateTime
 import org.balhom.currencyprofilesapi.common.utils.TestDataUtils.Companion.randomText
-import java.util.UUID
+import java.util.*
 
 class MockCurrencyProfileFactory {
     companion object {
@@ -22,10 +22,10 @@ class MockCurrencyProfileFactory {
                     .entries
                     .toTypedArray()
                     .random(),
-                balance = randomDouble(0.0, 10000.0),
+                balance = randomBigDecimal(0.0, 10000.0),
                 initDate = randomPastDateTime().toLocalDate(),
-                goalMonthlySaving = randomDouble(0.0, 1000.0),
-                goalYearlySaving = randomDouble(0.0, 12000.0),
+                goalMonthlySaving = randomBigDecimal(0.0, 1000.0),
+                goalYearlySaving = randomBigDecimal(0.0, 12000.0),
                 imageData = if (randomBool())
                     MockFileReferenceDataFactory.create(
                         currencyProfileId
